:rabbit: :koala: review

+++ src/sql/structure.sql
@@ -652,7 +652,7 @@ CREATE TABLE `items` (
   `book_users_can_in_past` TINYINT UNSIGNED NOT NULL DEFAULT 0,
   `book_is_cancellable` TINYINT UNSIGNED NOT NULL DEFAULT 1,
   `book_cancel_minutes` INT UNSIGNED NOT NULL DEFAULT 0,
-  `book_maximum_days_in_advance` INT UNSIGNED NOT NULL DEFAULT 0,
+  `book_maximum_days_in_advance` TINYINT UNSIGNED NOT NULL DEFAULT 0,
   `is_procurable` TINYINT UNSIGNED NOT NULL DEFAULT 0,
   `proc_pack_qty` MEDIUMINT UNSIGNED NOT NULL DEFAULT 0,
   `proc_price_notax` DECIMAL(10, 2) UNSIGNED NOT NULL DEFAULT 0.00,
